+++ src/index.ts
@@ -1,5 +1,8 @@
 export { Canvas } from "./Canvas";
 export { Brush } from "./Brush";
+export { Layer } from "./Layer";
+export { LayerManager } from "./LayerManager";
+export type { CreateLayerOptions, UpdateLayerOptions } from "./LayerManager";
 export * from "./utils";
 export * from "./types";
 export * from "./modules";